requirements.txt
Flask>=1.1.2
imageio>=2.8.0
matplotlib>=3.2.1
Pillow>=7.1.2
Pyrebase4>=4.3.0
gunicorn>=20.0.4
numpy>=1.18.4
opencv-python>=4.2.0.34